Change beer cost from decimal to integer

+++ db/migrate/20180502231843_create_beers.rb
@@ -2,7 +2,7 @@ class CreateBeers < ActiveRecord::Migration[5.2]
   def change
     create_table :beers do |t|
       t.string :name
-      t.decimal :cost
+      t.integer :cost
       t.string :color
       t.string :origin
       t.integer :flight_id
